:art: move messaging model definitions to own module
